refactor: agent-driven gws; Python is deterministic local-file docx engine
Per the principle that gws actions belong to the agent (flexible) and Python is
only for well-known, unchanging transforms:

- Delete lib/aaif_meetups/gws_cli.py and remove locate_tracker from tracker.py —
  no Drive calls in Python anymore.
- The 3 skills (event-status/create-event/update-event) now operate on a LOCAL
  docx path; each SKILL.md instructs the agent to run gws (locate/download/upload)
  directly and call the deterministic script in between.
- Revert create_chapter/create_series to their original self-contained state.
- office.py + tracker.py remain pure-Python docx surgery (read/clone/set-field/
  restamp + date math).
