+++ skills/aaif-create-event/SKILL.md
@@ -6,20 +6,45 @@ argument-hint: '<chapter|series> --title "..." --date "..."'
 
 # AAIF Create Event
 
-Clone the example event section in a chapter/series `Event Tracker.docx`, fill the
-detail block, and compute every phase task's DUE date backward from the event date.
-Mode is auto-detected: a chapter (under Chapters/) clones the in-person task set; an
-online series (under Online/) clones the online set. Aborts if the title already exists.
+Add a new event to a chapter/series `Event Tracker.docx`: clone the example event
+section, fill the detail block, and compute every phase task's DUE date backward from
+the event date (the template's exact cadence is preserved per task). Mode is implicit —
+a chapter clones the in-person task set, an online series the online set, because you
+download whichever tracker the folder holds.
 
-Prereq: the `gws` CLI must be installed and authenticated (`gws-cli-access`).
+**You (the agent) drive Google Drive via the `gws` CLI; the Python script only does the
+deterministic docx edit on a local file.** Prereq: `gws` installed and authenticated
+(`gws-cli-access`).
 
-## Run
+## Steps
 
-    python skills/aaif-create-event/scripts/create_event.py "<chapter|series>" \
-      --title "Eval Night · Builder Series" \
-      --date "Wed · August 12, 2026 · 18:00 — late" \
-      [--theme "..."] [--venue "..."] [--platform "..."] [--speakers "..."] \
-      [--luma "lu.ma/aaif-..."] [--capacity "..."] [--organizer "..."] [--dry-run]
+1. **Locate the tracker.** Chapters parent `1IQ1K7aVOKUUkxAcfLuNjdETEnmavvtjx`, Online
+   parent `1g2vHrqDHfh9wBkDJryJIl8wqXA4J-d4i`. Find the named folder, then its
+   `Event Tracker.docx` id (see `aaif-event-status` for the exact `gws drive files list`
+   queries).
 
-Anything you omit is left as the example's text for you to fill later. Due-dates
-keep the template's exact cadence (each task's offset from the event date is preserved).
+2. **Download it:**
+
+   ```
+   gws drive files get --params '{"fileId":"<DOC_ID>","alt":"media"}' --output tracker.docx
+   ```
+
+3. **Add the event (deterministic, local).** Aborts if the title already exists:
+
+   ```
+   python skills/aaif-create-event/scripts/create_event.py tracker.docx \
+     --title "Eval Night · Builder Series" \
+     --date "Wed · August 12, 2026 · 18:00 — late" \
+     [--theme ...] [--venue ...] [--platform ...] [--speakers ...] \
+     [--luma ...] [--capacity ...] [--organizer ...] [--dry-run]
+   ```
+   Omitted fields keep the example's text for the organizer to fill later.
+
+4. **Upload it back:**
+
+   ```
+   gws drive files update --params '{"fileId":"<DOC_ID>"}' --upload tracker.docx \
+     --upload-content-type application/vnd.openxmlformats-officedocument.wordprocessingml.document
+   ```
+
+Use `--dry-run` in step 3 first if you want to preview without modifying the local file.
