+++ skills/aaif-create-event/scripts/create_event.py
@@ -1,15 +1,14 @@
 #!/usr/bin/env python3
-"""Create a new event section in a chapter/series Event Tracker.docx: clone the
-example section, fill details, and stamp all phase due-dates from the event date.
-Stdlib-only, pure-Python docx edit; Drive I/O via the gws CLI."""
+"""Deterministic, local-file event creator: clone the example section in an
+Event Tracker.docx, fill details, and stamp all phase due-dates from the event date.
+Operates on a docx the agent has ALREADY downloaded via the gws CLI — this script
+never touches Drive. Pure-Python docx edit."""
 import argparse
-import os
 import pathlib
 import sys
-import tempfile
 
 sys.path.insert(0, str(pathlib.Path(__file__).resolve().parents[3] / "lib"))
-from aaif_meetups import gws_cli, office, tracker  # noqa: E402
+from aaif_meetups import office, tracker  # noqa: E402
 
 # CLI flag -> detail label
 FIELD_MAP = {"title": "EVENT TITLE", "date": "DATE & TIME", "theme": "THEME / SERIES",
@@ -34,8 +33,8 @@ def _fields_from_args(a):
 
 
 def main():
-    ap = argparse.ArgumentParser()
-    ap.add_argument("group", help="chapter or series name")
+    ap = argparse.ArgumentParser(description="Add an event to a local Event Tracker.docx")
+    ap.add_argument("docx", help="path to a tracker.docx already downloaded via gws")
     ap.add_argument("--title", required=True)
     ap.add_argument("--date", required=True, help='e.g. "Wed · August 12, 2026 · 18:00 — late"')
     for f in ("theme", "venue", "platform", "speakers", "luma", "capacity",
@@ -46,22 +45,16 @@ def main():
 
     event_date = tracker.parse_event_date(a.date)
     fields = _fields_from_args(a)
-    loc = tracker.locate_tracker(a.group)
-    print("Tracker: %s (%s)  event: %s  date: %s"
-          % (loc["folder_name"], loc["kind"], a.title, event_date))
+    root = office.read_document(a.docx)
+    if any(a.title.lower() in e["title"].lower() for e in tracker.list_events(root)):
+        sys.exit("ABORT: an event titled %r already exists in %s." % (a.title, a.docx))
+    print("Event: %s  date: %s  (fields set: %s)"
+          % (a.title, event_date, ", ".join(sorted(fields)) or "title/date only"))
     if a.dry_run:
         print("[dry-run] would clone the example section and stamp dates; no write.")
         return
-    with tempfile.TemporaryDirectory() as d:
-        path = os.path.join(d, "tracker.docx")
-        gws_cli.gws_download(loc["file_id"], path)
-        root = office.read_document(path)
-        if any(a.title.lower() in e["title"].lower() for e in tracker.list_events(root)):
-            sys.exit("ABORT: an event titled %r already exists in this tracker." % a.title)
-        tracker.add_event(root, fields, event_date)
-        office.save_document(path, root, path)
-        gws_cli.gws_upload(loc["file_id"], path, gws_cli.DOCX)
-    print("Done. New event section added and due-dates stamped.")
+    apply_local(a.docx, fields, event_date)
+    print("Done. New event section added to %s and due-dates stamped." % a.docx)
 
 
 if __name__ == "__main__":
